requirements.txt
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7
SQLAlchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
pillow==10.1.0
